Addresses some performance issues identified in #749.

Primary changes:
1. LpConstraint no longer inherits from LpAffineExpression, this is so a reference can be kept to the already created object instead of copying it
2. Avoid creating temporary LpAffineExpression when building a LpConstraint with a RHS that is an int/float
3. Avoid creating temporary lists (e.g., of Iterables before iterating over them, before transforming lists into another form)

In some experiments with a model I am working with I have observed 4x improvement in setting constraints. pulp/apis/cplex_api.py:buildSolverModel has gone from 70s to 50s.

Comparison of numbers in performance script from #749 comparing the original performance to performance with this PR. Using list of tuples and passing that directly to LpAffineExpression remains faster than using lpSum.

| Name      | Original      | PR |
| ------------- | ------------- | ------------- |
| Slow Objective | 16.692 | 16.595 |
| Slow Constraint | 22.000 | 6.435 |
| Fast Objective | 2.945 | 2.999 |
| Fast Constraint | 20.278 | 5.552 |