@MBradbury 2024-05-04
Just a note that I added some type hinting to help me understand the code. Happy to remove if needed.