@MBradbury 2024-05-06
> Thanks for the changes. they look comprehensive. I also appreciate the type hints.
> 
> I'm curious though if there was a way to keep the inheritance of `LpAffineExpression` without making a copy when initializing the `LpConstraint`. Can't we just modifiy the `LpConstraint` init in line 1012 to fill the values without copying them? and then we would keep everything else as is.

This will still lead to a large copy as it is essentially copying a dict to another dict. So for the test file that would be a copy of a 3 million entry dict.

Another approach could be to have LpAffineExpression not inherit from dict and instead have it as a variable. Then the constraint would just use that variable when inherited from LpAffineExpression, avoiding the large copy.