@stumitchell 2024-05-06
Looking good however, I suggest that we do a major version change as the change to modifying things in place instead of copying may cause errors in existing code.