@MBradbury 2024-05-27
Looks like `LpConstraint` needs a `asCplexLpAffineExpression` method that calls the method of the same name on `expr`. Or the test can be updated to use `LpProblem.objective.expr.asCplexLpAffineExpression` instead of `LpProblem.objective.asCplexLpAffineExpression`.