@christiansegercrantz 2024-11-26
What's the current situation with this PR, is it still being worked on? I'm currently experiencing a very similar situation where the solving itself is taking 1.7s but I accumulate 22s of time from LpSum.