@MBradbury 2024-11-26
There was a small change that I mentioned in an earlier message, I just don't have the time to address it currently. I expect that this will be an issue for me in March/April, when I need to use Pulp for an annual task, so will likely take a look again then.

When I originally took a look the biggest design issue is that `LpAffineExpression` inherits from a dict. Ideally this would not be the case to avoid excessive copying, but fixing this issue was even harder to get right than the current attempt at speeding up Pulp.