@christiansegercrantz 2025-02-19
Alright, so we are waiting for maintainer approval and release then?