@pchtsp 2025-02-19
thanks! can you please apply the black linter as per instructions here: https://coin-or.github.io/pulp/develop/contribute.html#applying-the-black-linter-formatter