@christiansegercrantz 2025-02-19
Great job! @pchtsp When can we expect a release with this?