@pchtsp 2025-02-19
@christiansegercrantz 
following @stumitchell advice above, I just release PuLP 3.0.0 to pypi with these changes.